requirements.txt
-e git://github.com/spiderware/python-spiderwaregps.git#egg=spiderwaregps
argparse==1.2.1
boto==2.6.0
certifi==0.0.8
chardet==2.1.1
cmsplugin-filer==0.9
coverage==3.5.3
dj-database-url==0.2.1
django-analytical==0.13.0
django-classy-tags==0.3.4.1
django-cms==2.3.3
django-dirtyfields==0.1
django-filer==0.9.1
django-mptt==0.5.2
django-polymorphic==0.2
django-pylibmc-sasl==0.2.4
django-redis-cache==0.9.5
django-secure==0.1.2
django-sekizai==0.6.1
django-setuptest==0.1.2
django-storages==1.1.5
Django==1.4.1
easy-thumbnails==1.1
gevent==0.13.8
gpxpy==0.7.2
greenlet==0.4.0
gunicorn==0.14.6
html5lib==0.95
ipdb==0.7
ipython==0.13
lxml==3.0
pep8==1.3.3
PIL==1.1.7
pip-tools==0.2
psycopg2==2.4.5
pykml==0.1.0
pylibmc==1.2.3
python-dateutil==2.1
raven==2.0.6
redis==2.7.1
requests==0.14.1
simplejson==2.6.2
six==1.2.0
South==0.7.6
spiderwaregps==0.3a1
